+++ CHANGES.md
@@ -4,6 +4,7 @@
   * NEW check: simple IPv6 Path MTU checks against RIPE Atlas anchors
   * CHANGED: move root DNS server config to the ini file
   * CHANGED: support "ping is expected to fail" per root server and AF
+  * CHANGED: all defined open DNS resolvers are now in one check
 
 ### 0.7.20260311
   * NEW check: SSH host key checks